requirement.txt
Django==1.8.5
MySQL-python==1.2.5
argparse==1.2.1
decorator==4.0.4
django-filter==0.11.0
djangorestframework==3.3.0
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
simplegeneric==0.8.1
traitlets==4.0.0
wsgiref==0.1.2
